requirements.txt
SQLAlchemy>=2.0
alembic
psycopg2-binary
python-dotenv
click
schedule
pytest
